Revert "Move all tests to a tests folder"
This reverts commit be60f3b34e78b4e92131b094be19cdc0580ae382.
